0.5.2 configure access to the endpoints

+++ Recipes/task/src/recipes/controller/RecipeController.java
@@ -3,6 +3,7 @@ package recipes.controller;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.server.ResponseStatusException;
@@ -11,6 +12,7 @@ import recipes.service.RecipeService;
 
 import javax.validation.Valid;
 import javax.validation.constraints.Min;
+import java.security.Principal;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
@@ -27,18 +29,20 @@ public class RecipeController {
         this.recipeService = recipeService;
     }
 
+    @PreAuthorize("isAuthenticated()")
     @PostMapping("new")
     @ResponseBody
-    public ResponseEntity<Map<String, Long>> createRecipe(@Valid @RequestBody Recipe recipe) {
+    public ResponseEntity<Map<String, Long>> createRecipe(@Valid @RequestBody Recipe recipe, Principal principal) {
         try {
-            long id = recipeService.saveRecipe(recipe);
+            String creator = principal.getName();
+            long id = recipeService.saveRecipe(recipe, creator);
             return new ResponseEntity<>(Map.of("id", id), HttpStatus.OK);
         } catch (Exception e) {
             return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
         }
     }
 
-
+    @PreAuthorize("isAuthenticated()")
     @GetMapping("search")
     @ResponseBody
     public ResponseEntity<List<Recipe>> searchRecipesByParams(@RequestParam Map<String, String> param) {
@@ -51,6 +55,8 @@ public class RecipeController {
         return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
     }
 
+
+    @PreAuthorize("isAuthenticated()")
     @GetMapping("{id}")
     public ResponseEntity<Recipe> getRecipe(@PathVariable @Min(1) long id) {
         Optional<Recipe> recipe = recipeService.findRecipeById(id);
@@ -60,17 +66,34 @@ public class RecipeController {
     }
 
 
+    @PreAuthorize("isAuthenticated()")
     @PutMapping("{id}")
-    public ResponseEntity<HttpStatus> updateRecipe(@PathVariable long id, @Valid @RequestBody Recipe newRecipe) {
-        boolean status = recipeService.updateRecipeById(id, newRecipe);
-        return status ? new ResponseEntity<>(HttpStatus.NO_CONTENT) : new ResponseEntity<>(HttpStatus.NOT_FOUND);
+    public ResponseEntity<HttpStatus> updateRecipe(@PathVariable long id,
+                                                   @Valid @RequestBody Recipe newRecipe,
+                                                   Principal principal) {
+        String creatorName = principal.getName();
+        Optional<Recipe> oldRecipe = recipeService.findRecipeById(id);
+
+        if (oldRecipe.isEmpty()) return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        if (!creatorName.equals(oldRecipe.get().getCreator())) return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+
+        recipeService.updateRecipeById(id, newRecipe);
+        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
     }
 
 
+    @PreAuthorize("isAuthenticated()")
     @DeleteMapping("/{id}")
-    public ResponseEntity<HttpStatus> deleteRecipe(@PathVariable @Min(1) long id) {
-        boolean status = recipeService.deleteRecipeById(id);
-        return status ? new ResponseEntity<>(HttpStatus.NO_CONTENT) : new ResponseEntity<>(HttpStatus.NOT_FOUND);
+    public ResponseEntity<HttpStatus> deleteRecipe(@PathVariable @Min(1) long id, Principal principal) {
+        String creatorName = principal.getName();
+
+        Optional<Recipe> delRecipe = recipeService.findRecipeById(id);
+
+        if (delRecipe.isEmpty()) return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        if (!creatorName.equals(delRecipe.get().getCreator())) return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+
+        recipeService.deleteRecipeById(id);
+        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
     }
 
 }
